Highlight :target background
Sadly this is not perfect, I would have really liked if this highlight
was slightly wrapped with 1rem of padding. But by default elements
typically have 1em of margin instead of padding. We are currently
relying on this default and I don't really want to change it right now,
especially not without a visual testing suite and having a better
understanding about the use of negative spacing in design
Holding back this one for later